[analytics] Added guard against invalid queue ids
