+++ packages/shared-dto/src/queue.ts
@@ -40,6 +40,7 @@ interface GETProgressRes {
   currPlace: number;
   total: number;
 }
+
 interface GETAllReq {
   queueId: string;
 }
