+++ packages/ui/src/components/dashboard/analytics/widgets/QueueListWidget.tsx
@@ -11,39 +11,11 @@ import {
   Th,
   Image,
 } from "@chakra-ui/react";
-import { useEffect, useState } from "react";
-import { useParams } from "react-router-dom";
-import { config } from "../../../../lib/config";
+
 // @ts-ignore
 import queuelist from "../../../media/analytics/queuelist.svg";
 
-const UsersTable = () => {
-  const [queuedUsers, setQueuedUsers] = useState([]);
-  const { queueId } = useParams();
-
-  // fetch users for this queue
-  const fetchUsers = async (queueId) => {
-    try {
-      const res = await fetch(
-        `${config.hostUrl}/api/queue/all?queueId=${queueId}`,
-        {
-          method: "GET",
-          cache: "no-cache",
-          credentials: "include",
-        },
-      );
-
-      const { users } = await res.json();
-      setQueuedUsers(users);
-    } catch (err) {
-      console.error(err.message);
-    }
-  };
-
-  useEffect(() => {
-    fetchUsers(queueId);
-  }, []);
-
+const UsersTable = ({ queuedUsers }) => {
   if (queuedUsers.length == 0) {
     return <Image boxSize="150" pb="5%" src={queuelist}></Image>;
   }
@@ -72,14 +44,14 @@ const UsersTable = () => {
   );
 };
 
-const QueueListWidget = () => {
+const QueueListWidget = ({ queuedUsers }) => {
   return (
     <Box boxShadow="xs" rounded="lg" bg="white" height="100%" w="90%">
       <VStack>
         <Heading pt="5%" pb="5%" fontSize={"2xl"}>
           Users
         </Heading>
-        <UsersTable></UsersTable>
+        <UsersTable queuedUsers={queuedUsers}></UsersTable>
       </VStack>
     </Box>
   );
